backend/pyproject.toml
[project]
name = "enlaight-backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "library (>=3.0.113,<4.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "django-filter (>=25.1,<26.0)",
    "i18n (>=0.2,<0.3)",
    "boto3 (>=1.40.39,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)"
]

[tool.poetry]
packages = [
    { include = "core", from = "src" },
    { include = "authentication", from = "src" }
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
